test folder organize, working to clean App.tsx
