+++ src/__tests__/Containers/VisualContainer.unit.test.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import "@testing-library/jest-dom";
 
 import { VisualContainer } from '../../app/Containers/VisualContainer';
-import { cleanup, generateStore, render, screen } from '../testing';
+import { cleanup, generateStore, render, screen } from '../testignore/testing';
 // this is our mock state that we will use to run our tests (for future tests)
 import { snapshotHistoryMock } from '../../../mock/state-snapshot'
