+++ src/__tests__/components/Metrics/__tests__/RankedGraph.unit.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RankedGraph from '../RankedGraph';
+import RankedGraph from '../../../../app/components/Metrics/RankedGraph';
 import {render, cleanup, screen} from '@testing-library/react';
 import '@testing-library/dom';
 import '@testing-library/jest-dom'
